backend/requirements.txt
Flask
flask-jwt-extended
flask-cors
python-dotenv
Flask-Limiter
